+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 ## Version 1.0.0-alpha.1
 - Initial alpha release
+## Version 1.0.0-alpha.2
+- Initial movement and interactions restrictions
+## Version 1.0.0-alpha.3
+- Initial register and login commands